+++ Wabbajack.App.Blazor/Components/TopBar.razor
@@ -28,7 +28,7 @@
     {
         {"Play", Play.Route},
         {"Gallery", Gallery.Route},
-        {"Install", Install.Route},
+        {"Install", Select.Route},
         {"Create", Create.Route}
     };
