+++ Wabbajack.App.Blazor/Pages/Install/Configure.razor
@@ -1,0 +1,48 @@
+﻿@page "/install/configure"
+
+@namespace Wabbajack.App.Blazor.Pages
+
+<div id="content">
+    <div class="install-background">
+        <img id="background-image" src="@ModlistImage" alt=""/>
+    </div>
+    <div class="list">
+        @if (Modlist is not null)
+        {
+            <div class="left-side">
+                <InfoBlock Title="@Modlist.Name" Subtitle="@Modlist.Author" Comment="@Modlist.Version.ToString()" Description="@Modlist.Description"/>
+            </div>
+            <div class="right-side">
+                <InfoImage Image="@ModlistImage"/>
+            </div>
+        }
+    </div>
+    <div class="settings">
+        <div class="locations">
+            <div class="labels">
+                <span>Target Modlist</span>
+                <span>Install Location</span>
+                <span>Download Location</span>
+            </div>
+            <div class="paths">
+                <span class="modlist-file">@ModlistPath.ToString()</span>
+                <span class="install-location" @onclick="SelectInstallFolder">@InstallPath.ToString()</span>
+                <span class="download-location" @onclick="SelectDownloadFolder">@DownloadPath.ToString()</span>
+
+            </div>
+        </div>
+        <div class="options">
+            <OptionCheckbox Label="Overwrite Installation"/>
+            <OptionCheckbox Label="NTFS Compression"/>
+            <OptionCheckbox Label="Do a sweet trick"/>
+            <OptionCheckbox Label="Something else"/>
+        </div>
+        <div class="install">
+            <img src="images/icons/play.svg" @onclick="Install" alt="Install">
+        </div>
+    </div>
+</div>
+
+@code {
+        public const string Route = "/install/configure";
+}
