+++ Wabbajack.App.Blazor/Pages/Install/Configure.razor.scss
@@ -1,4 +1,4 @@
-﻿@import "../Shared/Globals.scss";
+﻿@import "../../Shared/Globals.scss";
 
 $checkbox-background: rgba(255, 255, 255, 0.2);
 $checkbox-background-hover: darkgrey;
