+++ Wabbajack.App.Blazor/Pages/Install/Select.razor
@@ -1,4 +1,4 @@
-﻿@page "/install"
+﻿@page "/install/select"
 
 @namespace Wabbajack.App.Blazor.Pages
 
@@ -26,5 +26,5 @@
 </div>
 
 @code {
-    public const string Route = "/install";
+    public const string Route = "/install/select";
 }
