+++ Wabbajack.App.Blazor/Pages/Install/Select.razor.scss
